@@ -4,24 +4,40 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "production": "webpack -p",
-    "start": "webpack-dev-server --inline --hot --port 9000"
+    "build": "npm run clean && webpack -p",
+    "start": "webpack-dev-server --no-info --hot --inline --port 9000 --open",
+    "clean": "rimraf ./dist && mkdir dist",
+    "lint": "eslint ./app --ext .js --env browser --color"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.12.0",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "react-router": "^2.0.0"
+    "axios": "0.14.0",
+    "classnames": "2.2.5",
+    "normalize.css": "4.2.0",
+    "react": "15.3.1",
+    "react-dom": "15.3.1",
+    "react-router": "2.7.0"
   },
   "devDependencies": {
     "babel-core": "^6.7.0",
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "css-loader": "^0.25.0",
+    "eslint": "^3.4.0",
+    "eslint-config-airbnb": "^10.0.1",
+    "eslint-plugin-import": "^1.14.0",
+    "eslint-plugin-jsx-a11y": "^2.2.1",
+    "eslint-plugin-react": "^6.2.0",
+    "file-loader": "^0.9.0",
     "html-webpack-plugin": "^2.9.0",
-    "webpack": "^1.12.14",
-    "webpack-dev-server": "^1.14.1"
+    "postcss-cssnext": "^2.8.0",
+    "postcss-loader": "^0.11.1",
+    "rimraf": "^2.5.4",
+    "style-loader": "^0.13.0",
+    "url-loader": "^0.5.7",
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.15.1"
   }
 }
