@@ -1,12 +1,12 @@
 import React from 'react';
-import {ReactRouter, Router, Route, IndexRoute, browserHistory} from 'react-router';
+import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import Main from '../components/main';
 import Home from '../components/home';
 
-var routes = (
+const routes = (
 	<Router history={browserHistory}>
-		<Route path='/' component={Main}>
-			<IndexRoute component={Home}></IndexRoute>
+		<Route path="/" component={Main}>
+			<IndexRoute component={Home} />
 		</Route>
 	</Router>
 );
