@@ -1,8 +1,7 @@
-import React from 'react';
 import ReactDom from 'react-dom';
 import routes from './config/routes';
 
 ReactDom.render(
-	routes, 
+	routes,
 	document.getElementById('app')
 );
